Split out integration tests
Until we add a way to exclude tags, it's not really practical to have
the integration tests in the same file as the real conformance tests.

Toil's CI wants to run them separately, and it also can only address the
tests to skip over by number.

If we have the integration tests in with the conformance tests, we want
to keep tham at the end, and insert new conformance tests before them.
But that makes their numbers change, so Toil has no good way for its CI
to exclude them from the conformance job.

If we put them in a separate file like this, and append conformance
tests, we can keep track of the known failing-in-Toil tests by number
and not run them in CI, without worrying about numbers constantly
shifting.