Extract output types manually better and more
This uses the regex-based parser whenever anything goes wrong with
MiniWDL parsing, not just when MiniWDL complains that Object is used.

It also adjusts the regex-based parser to look just at the output
section, within the task/workflow. That part was written with the "help"
of Anthropic Claude, and then substantially rewritten by me because I
didn't like it.

With this and https://github.com/openwdl/wdl/pull/731 I'm able to import
the WDL spec tests.